fix header / file structure
